% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/const.R
\name{const}
\alias{const}
\title{Constructor Function}
\usage{
const(
  plotType = "scatterplot",
  cor = FALSE,
  mile = TRUE,
  length = TRUE,
  weight = TRUE,
  ddt = TRUE,
  FCM = TRUE,
  LCM = TRUE,
  SCM = TRUE,
  TRM = TRUE,
  catfish = TRUE,
  buffalo = TRUE,
  bass = TRUE,
  swap = FALSE
)
}
\arguments{
\item{plotType}{the type of plot to be made}

\item{cor}{whether or not to use the correlation matrix for PCA}

\item{mile}{whether or not to include the mile variable}

\item{length}{whether or not to include the length variable}

\item{weight}{whether or not to include the weight variable}

\item{ddt}{whether or not to include the DDT variable}

\item{FCM}{whether or not to include the FCM River}

\item{LCM}{whether or not to include the LCM River}

\item{SCM}{whether or not to include the SCM River}

\item{TRM}{whether or not to include the TRM River}

\item{catfish}{whether or not to include the CCATFISH species}

\item{buffalo}{whether or not to include the SMBUFFALO species}

\item{bass}{whether or not to include the LMBASS species}
}
\value{

}
\description{
Constructor Function
}
