% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT Data Set}
\format{a list with 6 variables, which are:
\describe{
\item{MILE}{ the mile of the river at which the fish was taken}
\item{LENGTH}{ the length of the wish}
\item{WEIGHT}{ the weight of the fish}
\item{DDT}{ the DDT concentration in the fish}
\item{RIVER}{ the river from which the fish was sampled}
\item{SPECIES}{ the species of fish}
}}
\usage{
ddt
}
\description{
A data set collected by the US forces from the Tennessee River and its tributaries
}
\keyword{datasets}
