% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter}
\alias{plotter}
\title{Shiny Plotting Function}
\usage{
plotter(
  plotType = "screeplot",
  mile = TRUE,
  length = TRUE,
  weight = TRUE,
  ddt = TRUE,
  FCM = TRUE,
  LCM = TRUE,
  SCM = TRUE,
  TRM = TRUE,
  catfish = TRUE,
  buffalo = TRUE,
  bass = TRUE,
  pcaType = "covariance",
  col = "RIVER"
)
}
\arguments{
\item{plotType}{the type of plot to be made}

\item{mile}{whether or not to include the MILE variable}

\item{length}{whether or not to include the LENGTH variable}

\item{weight}{whether or not to include the WEIGHT variable}

\item{ddt}{whether or not to include the DDT variable}

\item{FCM}{whether or not to include the FCM River}

\item{LCM}{whether or not to include the LCM River}

\item{SCM}{whether or not to include the SCM River}

\item{TRM}{whether or not to include the TRM River}

\item{catfish}{whether or not to include the CCATFISH Species}

\item{buffalo}{whether or not to include the SMBUFFALO Species}

\item{bass}{whether or not to includ ethe LMBASS Species}

\item{pcaType}{whether to use the covariance or correlation matrix for PCA}

\item{col}{variable to use to differentiate points in scatterplot}
}
\value{
a plot of data
}
\description{
Shiny Plotting Function
}
\examples{
plotter("scatterplot",length=FALSE,buffalo=FALSE)
}
